+++ src/components/CustomWeb3PostEmbed/FarCasterPost.js
@@ -14,17 +14,67 @@ import { useRouter } from 'next/router';
 import { TruncateText } from '../styles';
 import { useFarcasterReplyParent } from '../../hooks/queries';
 import withSuspense from '../../hoc/withSuspense';
+import Reply from './Reply';
+import Avatar from './Avatar';
+import HeaderSection from './HeaderSection';
 
-const FarCasterPost = ({ post, text, postid, attachments }) => {  
+const FarCasterPost = ({ post, text, postid, attachments, classes }) => {  
   const { pathname } = useRouter();
-  const replyParent = useFarcasterReplyParent(post?.meta?.replyParentMerkleRoot)
+  const web3Preview = post
+  const {id} = post
+ // const replyParent = useFarcasterReplyParent(post?.meta?.replyParentMerkleRoot)
   let parsedText = parseText(text);
+  const parents = post.meta.parents
+  const isReply = parents?.length  > 0
+  // const isReplyToReply = parents.length > 1
+  
   const isFullPost = () => {
     return pathname === '/post/[id]';
   };
-
+  if(isReply){
+    return (
+      <Grid item="item" xs={12}>
+        <Reply classes={classes} post={post}>
+          </Reply>
+          </Grid>)
+  }
+  // else if(isReplyToReply){
+  //   return <Reply></ReplyToReply>
+  // } 
+  else {
   return ( 
-  <Link href={`/post/${postid}`} >
+    <Grid item="item" xs={12}>
+        <Grid container="container" direction="row" spacing={1}>
+          <Grid item="item">
+            <Avatar
+              classes={classes}
+              url={web3Preview.creator.avatarUrl}
+              tweetLink={id}
+            />
+          </Grid>
+          <Grid item="item" xs>
+            <Grid container="container" direction="row" spacing={0}>
+              <Grid item="item" xs={12}>
+                <HeaderSection
+                  classes={classes}
+                  name={web3Preview.creator.fullname}
+                  handle={web3Preview.creator.handle}
+                  address={web3Preview.creator.address}
+                  protocol={web3Preview.protocol}
+                  replyParentUsername={
+                    web3Preview.meta.replyParentUsername?.username
+                  }
+                  tweetLink={id}
+                  createdAt={web3Preview.createdAt}
+                />
+              </Grid>
+              <Grid item="item" xs={12}>
+                <Grid container="container" spacing={1}>
+                  {postid}
+                  <Grid item="item" xs={12}>
+                    {/* <Link href={tweetLink} target="_blank" underline="none"> */}
+                    <Typography variant="body2">
+                    <Link href={`/post/${postid}`} >
       <Grid
         item
         sx={{cursor:'pointer'}}>
@@ -75,7 +125,19 @@ const FarCasterPost = ({ post, text, postid, attachments }) => {
       </Grid>
       
       </Link>
+                    </Typography>
+                    {/* </Link> */}
+                  </Grid>
+                </Grid>
+              </Grid>
+            </Grid>
+          </Grid>
+        </Grid>
+      </Grid>
+  
   );
+    
+}
 };
 
 export default withSuspense()(FarCasterPost);
