+++ src/components/CustomWeb3PostEmbed/Reply.js
@@ -9,132 +9,85 @@ import Avatar from './Avatar';
 
 const DEFAULT_TWITTER_PROF = '/images/default-twitter-prof.png';
 
-const Reply = ({ tweetData, classes }) => {
-  const { user } = tweetData.tweetInfo;
-  const { url } = tweetData;
-  const [previewData, setPreviewData] = useState(null);
-  const entities = tweetData.tweetInfo.entities
-    ? tweetData.tweetInfo.entities
-    : false;
-  const entitiesURLS = entities && entities.urls.length > 0;
-  const tweetLink = url || '';
+const Reply = ({ post,  classes }) => {
+  const parents = post.meta.parents;
+  const userName = post.creator.fullname;
+  const userHandle = post.creator.handle;
+  const userAvatar = post.meta.avatar;
+  const userPostText = parseText(post.content);
+  const userPostLink = `farcaster://profiles/${post.creator.address}/posts`;
 
-  // ORIGINAL
-  const extendedEntities = tweetData.tweetInfo.extended_entities
-    ? tweetData.tweetInfo.extended_entities
-    : false;
+  const directParent = parents[0];
+  const directParentName = directParent.meta.displayName;
+  const directParentHandle = directParent.body.username;
+  const directParentAvatar = directParent.meta.avatar;
+  const directParentPostText = parseText(directParent.body.data.text);
+  const directParentPostLink = `farcaster://profiles/${directParent.body.address}/posts`;
+console.log({directParentPostLink,userPostLink})
+  // const { url } = tweetData;
+  // const [previewData, setPreviewData] = useState(null);
 
-  useEffect(() => {
-    if (entitiesURLS) {
-      if (entities.urls[0].expanded_url) {
-        (async () => {
-          try {
-            const previewData = await fetchLinkPreviewData(
-              entities.urls[0].expanded_url
-            );
-            setPreviewData(previewData);
-          } catch (e) {
-            console.log(e);
-          }
-        })();
-      }
-    }
-  }, []);
-
-  let hasMedia;
-  if (extendedEntities) {
-    hasMedia = extendedEntities.media
-      ? extendedEntities.media.length > 0
-      : false;
-  }
-
-  let mediaURL, mediaType, hasPhoto, hasVideo;
-  if (hasMedia) {
-    mediaURL = extendedEntities.media[0].media_url_https
-      ? extendedEntities.media[0].media_url_https
-      : false;
-    mediaType = extendedEntities.media[0].type;
-    hasPhoto = Boolean(mediaType === 'photo');
-    hasVideo = Boolean(mediaType === 'video' || mediaType === 'animated_gif');
-
-    if (hasVideo)
-      mediaURL = extendedEntities.media[0].video_info.variants[0].url;
-  }
-
-  let initialText;
-  if (tweetData.tweetInfo.text) {
-    initialText = tweetData.tweetInfo.text;
-  } else if (tweetData.tweetInfo.full_text) {
-    initialText = tweetData.tweetInfo.full_text;
-  } else {
-    initialText = '';
-  }
-
-  let tweetText = parseText(initialText);
 
   // REPLYS
-  let replyExtendedEntities =
-    tweetData.tweetInfo.reply_status &&
-    tweetData.tweetInfo.reply_status.extended_entities
-      ? tweetData.tweetInfo.reply_status.extended_entities
-      : false;
-  let replyHasMedia;
-  if (replyExtendedEntities) {
-    replyHasMedia = replyExtendedEntities.media
-      ? replyExtendedEntities.media.length > 0
-      : false;
-  }
+  // let replyExtendedEntities =
+  //   tweetData.tweetInfo.reply_status &&
+  //   tweetData.tweetInfo.reply_status.extended_entities
+  //     ? tweetData.tweetInfo.reply_status.extended_entities
+  //     : false;
+  // let replyHasMedia;
+  // if (replyExtendedEntities) {
+  //   replyHasMedia = replyExtendedEntities.media
+  //     ? replyExtendedEntities.media.length > 0
+  //     : false;
+  // }
 
-  let replyMediaURL, replyMediaType, replyHasPhoto, replyHasVideo;
-  if (replyHasMedia) {
-    replyMediaURL = replyExtendedEntities.media[0].media_url_https
-      ? replyExtendedEntities.media[0].media_url_https
-      : false;
-    replyMediaType = replyExtendedEntities.media[0].type;
-    replyHasPhoto = Boolean(replyMediaType === 'photo');
-    replyHasVideo = Boolean(
-      replyMediaType === 'video' || replyMediaType === 'animated_gif'
-    );
+  // let replyMediaURL, replyMediaType, replyHasPhoto, replyHasVideo;
+  // if (replyHasMedia) {
+  //   replyMediaURL = replyExtendedEntities.media[0].media_url_https
+  //     ? replyExtendedEntities.media[0].media_url_https
+  //     : false;
+  //   replyMediaType = replyExtendedEntities.media[0].type;
+  //   replyHasPhoto = Boolean(replyMediaType === 'photo');
+  //   replyHasVideo = Boolean(
+  //     replyMediaType === 'video' || replyMediaType === 'animated_gif'
+  //   );
 
-    if (replyHasVideo)
-      replyMediaURL = replyExtendedEntities.media[0].video_info.variants[0].url;
-  }
+  //   if (replyHasVideo)
+  //     replyMediaURL = replyExtendedEntities.media[0].video_info.variants[0].url;
+  // }
 
-  let replyStatusText;
-  if (
-    tweetData.tweetInfo.reply_status &&
-    tweetData.tweetInfo.reply_status.text
-  ) {
-    replyStatusText = tweetData.tweetInfo.reply_status.text;
-  } else if (
-    tweetData.tweetInfo.reply_status &&
-    tweetData.tweetInfo.reply_status.full_text
-  ) {
-    replyStatusText = tweetData.tweetInfo.reply_status.full_text;
-  } else {
-    replyStatusText = '';
-  }
+  // let replyStatusText;
+  // if (
+  //   tweetData.tweetInfo.reply_status &&
+  //   tweetData.tweetInfo.reply_status.text
+  // ) {
+  //   replyStatusText = tweetData.tweetInfo.reply_status.text;
+  // } else if (
+  //   tweetData.tweetInfo.reply_status &&
+  //   tweetData.tweetInfo.reply_status.full_text
+  // ) {
+  //   replyStatusText = tweetData.tweetInfo.reply_status.full_text;
+  // } else {
+  //   replyStatusText = '';
+  // }
 
-  let text = parseText(replyStatusText);
-  let replyTweetText = text.split(' ').map((string) => linkMentions(string));
+  // let text = parseText(replyStatusText);
+  // let replyTweetText = text.split(' ').map((string) => linkMentions(string));
 
   // REPLY CUSTOM HEADER STYLING CONFIG
-  const userAvatar = classes.userAvatar;
-  const twitterName = classes.replyTwitterName;
-  const twitterBirdIcon = classes.twitterBirdIcon;
 
-  let replyScreenName, replyUserAvatar, replyName;
-  if (tweetData.tweetInfo.reply_status) {
-    replyScreenName = tweetData.tweetInfo.reply_status.user.screen_name;
-    replyUserAvatar =
-      tweetData.tweetInfo.reply_status.user.profile_image_url_https;
-    replyName = tweetData.tweetInfo.reply_status.user.name;
-  }
+  // let replyScreenName, replyUserAvatar, replyName;
+  // if (tweetData.tweetInfo.reply_status) {
+  //   replyScreenName = tweetData.tweetInfo.reply_status.user.screen_name;
+  //   replyUserAvatar =
+  //     tweetData.tweetInfo.reply_status.user.profile_image_url_https;
+  //   replyName = tweetData.tweetInfo.reply_status.user.name;
+  // }
 
-  const accountLink = `https://twitter.com/${replyScreenName}`;
-  const BothHaveMedia = hasMedia && replyHasMedia;
-  const smallImage = { maxHeight: 200, maxWidth: 550 };
-  const bigImage = { minHeight: 300, maxHeight: 400 };
+  // const accountLink = `https://twitter.com/${replyScreenName}`;
+  // const BothHaveMedia = hasMedia && replyHasMedia;
+  // const smallImage = { maxHeight: 200, maxWidth: 550 };
+  // const bigImage = { minHeight: 300, maxHeight: 400 };
 
   const addDefaultSrc = (e) => {
     e.target.onerror = null;
@@ -166,18 +119,7 @@ const Reply = ({ tweetData, classes }) => {
               spacing={1}
             >
               <Grid item>
-                {replyUserAvatar ? (
-                  <img
-                    className={userAvatar}
-                    src={replyUserAvatar}
-                    alt="user image"
-                    onError={addDefaultSrc}
-                  />
-                ) : (
-                  <span className={classes.letterAvatar}>
-                    {replyName && replyName[0] && replyName[0].toUpperCase()}
-                  </span>
-                )}
+              <Avatar classes={classes} url={directParentAvatar}  />
               </Grid>
               <Grid item xs>
                 <Grid container direction="row" style={{ height: '100%' }}>
@@ -195,62 +137,28 @@ const Reply = ({ tweetData, classes }) => {
           </Grid>
           <Grid item xs>
             <Grid container direction="column" spacing={1}>
-              <Grid item className={classes.header}>
-                <Grid
-                  container
-                  justifyContent="space-between"
-                  alignItems="flex-end"
-                >
-                  <Grid item>
-                    <Grid
-                      container
-                      spacing={1}
-                      direction="row"
-                      className={classes.userTags}
-                    >
-                      <Grid item>
-                        <Link
-                          href={accountLink}
-                          target="_blank"
-                          underline="none"
-                        >
-                          <h4 className={twitterName}>{replyName}</h4>
-                        </Link>
-                      </Grid>
-                      <Grid item>
-                        <Link
-                          href={accountLink}
-                          target="_blank"
-                          underline="none"
-                        >
-                          <Grid className={classes.twitterHandle}>
-                            @{replyScreenName}
-                          </Grid>
-                        </Link>
-                      </Grid>
-                    </Grid>
-                  </Grid>
-                  <Grid item className={twitterBirdIcon}>
-                    <Link href={tweetLink} target="_blank" underline="none">
-                      <img
-                        src="/images/icons/twitter.svg"
-                        height="24"
-                        alt="twitter"
-                      />
-                    </Link>
-                  </Grid>
-                </Grid>
-              </Grid>
+            <HeaderSection
+                  classes={classes}
+                  name={directParentName}
+                  handle={directParentHandle}
+                  address={directParent.body.address}
+                  protocol={post.protocol}
+                  //WRONG, NEEDS TO BE PARENTS POST ID; NOT REPLY ID
+                  tweetLink={post.id}
+                  createdAt={directParent.body.publishedAt}
+                  noBird
+                />
+           
               <Grid
                 item
                 className={classes.replyContent}
                 style={{ paddingBottom: '8px' }}
               >
-                <Link href={tweetLink} target="_blank" underline="none">
-                  <Typography variant="body2">{replyTweetText}</Typography>
-                </Link>
+                {/* <Link  target="_blank" underline="none"> */}
+                  <Typography variant="body2">{directParentPostText}</Typography>
+                {/* </Link> */}
               </Grid>
-              {replyHasPhoto && replyMediaURL ? (
+              {/* {replyHasPhoto && replyMediaURL ? (
                 <Grid item className={classes.replyImageContainer}>
                   <img
                     className={classes.tweetImg}
@@ -270,7 +178,7 @@ const Reply = ({ tweetData, classes }) => {
                     <TweetVidPlayer url={replyMediaURL} />
                   </Grid>
                 )
-              )}
+              )} */}
             </Grid>
           </Grid>
         </Grid>
@@ -281,24 +189,27 @@ const Reply = ({ tweetData, classes }) => {
           <Grid item="item" xs={12}>
             <Grid container="container" direction="row" spacing={1}>
               <Grid item="item">
-                <Avatar classes={classes} user={user} tweetLink={tweetLink} />
+                <Avatar classes={classes} url={userAvatar}  />
               </Grid>
               <Grid item="item" xs>
                 <Grid container="container" direction="column" spacing={1}>
                   <Grid item="item">
-                    <HeaderSection classes={classes} user={user} hideBird />
+                    <HeaderSection classes={classes} name={userName} handle={userHandle} address={post.creator.address} 
+                  tweetLink={post.id}
+                  createdAt={post.createdAt}
+                  hideBird />
                   </Grid>
                   <Grid item>
-                    <Link href={tweetLink} target="_blank" underline="none">
+                    {/* <Link  target="_blank" underline="none"> */}
                       <Typography
                         variant="subtitle2"
                         className={classes.tweetText}
                       >
-                        {tweetText.replace(/@\S+\s?/gm, '')}
+                        {userPostText}
                       </Typography>
-                    </Link>
+                    {/* </Link> */}
                   </Grid>
-                  <Grid item>
+                  {/* <Grid item>
                     {previewData && !replyHasMedia && !mediaURL && (
                       <Grid>
                         <LinkPreview
@@ -326,7 +237,7 @@ const Reply = ({ tweetData, classes }) => {
                     ) : (
                       hasVideo && mediaURL && <TweetVidPlayer url={mediaURL} />
                     )}
-                  </Grid>
+                  </Grid> */}
                 </Grid>
               </Grid>
             </Grid>
